+++ src/iqm/iqm_client/iqm_client.py
@@ -17,6 +17,7 @@ Client for connecting to the IQM quantum computer server interface.
 
 from __future__ import annotations
 
+from collections.abc import Iterable
 from datetime import datetime
 from importlib.metadata import version
 import itertools
@@ -220,19 +221,19 @@ class IQMClient:
 
         for i, circuit in enumerate(circuits):
             try:
-                # validate the circuit against the local information in iqm.iqm_client.models._SUPPORTED_OPERATIONS
+                # validate the circuit against the static information in iqm.iqm_client.models._SUPPORTED_OPERATIONS
                 validate_circuit(circuit)
             except ValueError as e:
                 raise CircuitValidationError(f'The circuit at index {i} failed the validation').with_traceback(
                     e.__traceback__
                 )
 
-        architecture = self.get_quantum_architecture()
+        architecture = self.get_dynamic_quantum_architecture(calibration_set_id)
 
         self._validate_qubit_mapping(architecture, circuits, qubit_mapping)
         serialized_qubit_mapping = serialize_qubit_mapping(qubit_mapping) if qubit_mapping else None
 
-        # validate the circuit against the quantum architecture of the server
+        # validate the circuit against the calibration-dependent dynamic quantum architecture
         self._validate_circuit_instructions(
             architecture, circuits, qubit_mapping, validate_moves=options.move_gate_validation
         )
@@ -299,7 +300,7 @@ class IQMClient:
 
     @staticmethod
     def _validate_qubit_mapping(
-        architecture: QuantumArchitectureSpecification,
+        architecture: DynamicQuantumArchitecture,
         circuits: CircuitBatch,
         qubit_mapping: Optional[dict[str, str]] = None,
     ) -> None:
@@ -334,12 +335,12 @@ class IQMClient:
 
         # check that each mapped qubit is defined in the quantum architecture
         for _logical, physical in qubit_mapping.items():
-            if physical not in architecture.qubits:
-                raise CircuitValidationError(f'Qubit {physical} not present in quantum architecture')
+            if physical not in architecture.components:
+                raise CircuitValidationError(f'Component {physical} not present in dynamic quantum architecture')
 
     @staticmethod
     def _validate_circuit_instructions(
-        architecture: QuantumArchitectureSpecification,
+        architecture: DynamicQuantumArchitecture,
         circuits: CircuitBatch,
         qubit_mapping: Optional[dict[str, str]] = None,
         validate_moves: MoveGateValidationMode = MoveGateValidationMode.STRICT,
@@ -371,11 +372,13 @@ class IQMClient:
 
     @staticmethod
     def _validate_instruction(
-        architecture: QuantumArchitectureSpecification,
+        architecture: DynamicQuantumArchitecture,
         instruction: Instruction,
         qubit_mapping: Optional[dict[str, str]] = None,
     ) -> None:
-        """Check that the instruction targets a valid qubit locus in the given quantum architecture.
+        """Validate an instruction against the dynamic quantum quantum architecture.
+
+        Checks that the instruction uses a valid implementation, and targets a valid locus.
 
         Args:
           architecture: quantum architecture to check against
@@ -386,53 +389,72 @@ class IQMClient:
         Raises:
             CircuitValidationError: validation failed
         """
-        if instruction.name not in architecture.operations:
-            raise CircuitValidationError(
-                f"Instruction '{instruction.name}' is not supported by the quantum architecture."
-            )
-        allowed_loci = architecture.operations[instruction.name]
-        op_info = _SUPPORTED_OPERATIONS[instruction.name]
+        op_info = _SUPPORTED_OPERATIONS.get(instruction.name)
+        if op_info is None:
+            raise CircuitValidationError(f"Unknown quantum operation '{instruction.name}'.")
+
         # apply the qubit mapping if any
-        qubits = [qubit_mapping[q] for q in instruction.qubits] if qubit_mapping else list(instruction.qubits)
+        mapped_qubits = tuple(qubit_mapping[q] for q in instruction.qubits) if qubit_mapping else instruction.qubits
 
-        if op_info.allow_all_loci:
+        def check_locus_components(allowed_components: Iterable[str], msg: str) -> None:
+            """Checks that the instruction locus consists of the allowed components only."""
+            for q, mapped_q in zip(instruction.qubits, mapped_qubits):
+                if mapped_q not in allowed_components:
+                    raise CircuitValidationError(
+                        f'{instruction!r}: Component {q} = {mapped_q} {msg}.'
+                        if qubit_mapping
+                        else f'{instruction!r}: Component {q} {msg}.'
+                    )
+
+        if op_info.no_calibration_needed:
             # all QPU loci are allowed
-            for q, mapped_q in zip(instruction.qubits, qubits):
-                if mapped_q not in architecture.qubits:
-                    raise CircuitValidationError(
-                        f'{instruction}: Qubit {q} = {mapped_q} does not exist on the QPU.'
-                        if qubit_mapping
-                        else f'{instruction}: Qubit {q} does not exist on the QPU.'
-                    )
-            return
-        if op_info.factorizable:
-            # Check that all qubits in the locus are allowed by the architecture
-            allowed_qubits = set(q for locus in allowed_loci for q in locus)
-            for q, mapped_q in zip(instruction.qubits, qubits):
-                if mapped_q not in allowed_qubits:
-                    raise CircuitValidationError(
-                        f'Qubit {q} = {mapped_q} is not allowed as locus for {instruction.name}'
-                        if qubit_mapping
-                        else f'Qubit {q} is not allowed as locus for {instruction.name}'
-                    )
+            check_locus_components(architecture.components, msg='does not exist on the QPU')
             return
 
-        # Check that locus matches one of the loci defined in architecture
+        gate_info = architecture.gates.get(instruction.name)
+        if gate_info is None:
+            raise CircuitValidationError(
+                f"Operation '{instruction.name}' is not supported by the dynamic quantum architecture."
+            )
+
+        if instruction.implementation is not None:
+            # specific implementation requested
+            impl_info = gate_info.implementations.get(instruction.implementation)
+            if impl_info is None:
+                raise CircuitValidationError(
+                    f"Operation '{instruction.name}' implementation '{instruction.implementation}' "
+                    f'is not supported by the dynamic quantum architecture.'
+                )
+            allowed_loci = impl_info.loci
+            instruction_name = f'{instruction.name}.{instruction.implementation}'
+        else:
+            # any implementation is fine
+            allowed_loci = gate_info.loci
+            instruction_name = f'{instruction.name}'
+
+        if op_info.factorizable:
+            # Check that all the locus components are allowed by the architecture
+            check_locus_components(
+                set(q for locus in allowed_loci for q in locus), msg=f"is not allowed as locus for '{instruction_name}'"
+            )
+            return
+
+        # Check that locus matches one of the allowed loci
         all_loci = (
-            [list(x) for locus in allowed_loci for x in itertools.permutations(locus)]
+            tuple(tuple(x) for locus in allowed_loci for x in itertools.permutations(locus))
             if op_info.symmetric
             else allowed_loci
         )
-        if qubits not in all_loci:
+        if mapped_qubits not in all_loci:
             raise CircuitValidationError(
-                f'{instruction.qubits} = {tuple(qubits)} not allowed as locus for {instruction.name}'
+                f"{instruction.qubits} = {tuple(mapped_qubits)} is not allowed as locus for '{instruction_name}'"
                 if qubit_mapping
-                else f'{instruction.qubits} not allowed as locus for {instruction.name}'
+                else f"'{instruction.qubits} is not allowed as locus for '{instruction_name}'"
             )
 
     @staticmethod
     def _validate_circuit_moves(
-        architecture: QuantumArchitectureSpecification,
+        architecture: DynamicQuantumArchitecture,
         circuit: Circuit,
         qubit_mapping: Optional[dict[str, str]] = None,
         validate_moves: MoveGateValidationMode = MoveGateValidationMode.STRICT,
@@ -453,7 +475,7 @@ class IQMClient:
             return
         move_gate = 'move'
         # Check if MOVE gates are allowed on this architecture
-        if move_gate not in architecture.operations:
+        if move_gate not in architecture.gates:
             if any(i.name == move_gate for i in circuit.instructions):
                 raise CircuitValidationError('MOVE instruction is not supported by the given device architecture.')
             return
@@ -463,9 +485,8 @@ class IQMClient:
         if validate_moves == MoveGateValidationMode.ALLOW_PRX:
             allowed_gates.add('prx')
 
-        # TODO use architecture.computational_resonators when available instead of relying on COMP_R prefix.
-        all_resonators = {q for q in architecture.qubits if q.startswith('COMP_R')}
-        all_qubits = set(architecture.qubits) - all_resonators
+        all_resonators = set(architecture.computational_resonators)
+        all_qubits = set(architecture.qubits)
         if qubit_mapping:
             reverse_mapping = {phys: log for log, phys in qubit_mapping.items()}
             all_resonators = {reverse_mapping[q] if q in reverse_mapping else q for q in all_resonators}