+++ src/iqm/iqm_client/models.py
@@ -17,6 +17,8 @@ from __future__ import annotations
 
 from dataclasses import dataclass, field
 from enum import Enum
+from functools import cached_property
+import re
 from typing import Any, Final, Optional, Union
 from uuid import UUID
 
@@ -48,7 +50,7 @@ class NativeOperation:
     factorizable: bool = False
     """Iff True, any multi-component instance of this operation can be broken down to
     single-component instances, and calibration data is specific to single-component loci."""
-    allow_all_loci: bool = False
+    no_calibration_needed: bool = False
     """Iff true, the operation is always allowed on all QPU loci regardless of calibration state.
     Typically a metaoperation like barrier."""
 
@@ -56,7 +58,7 @@ class NativeOperation:
 _SUPPORTED_OPERATIONS: dict[str, NativeOperation] = {
     op.name: op
     for op in [
-        NativeOperation('barrier', 0, symmetric=True, allow_all_loci=True),
+        NativeOperation('barrier', 0, symmetric=True, no_calibration_needed=True),
         NativeOperation('measure', 0, {'key': (str,)}, {'feedback_key': (str,)}, factorizable=True),
         NativeOperation(
             'prx',
@@ -72,7 +74,8 @@ _SUPPORTED_OPERATIONS: dict[str, NativeOperation] = {
             {
                 'angle_t': (float, int),
                 'phase_t': (float, int),
-                'feedback_label': (str,),
+                'feedback_key': (str,),
+                'feedback_qubit': (str,),
             },
         ),
         NativeOperation('cz', 2, symmetric=True),
@@ -99,7 +102,8 @@ class Instruction(BaseModel):
     measure          >= 1        ``key: str``, ``feedback_key: str``     Measurement in the Z basis.
     prx              1           ``angle_t: float``, ``phase_t: float``  Phased x-rotation gate.
     cc_prx           1           ``angle_t: float``, ``phase_t: float``,
-                                 ``feedback_label: str``                 Classically controlled PRX gate.
+                                 ``feedback_qubit: str``,
+                                 ``feedback_key: str``                   Classically controlled PRX gate.
     cz               2                                                   Controlled-Z gate.
     move             2                                                   Moves a qubit state between a qubit and a
                                                                          computational resonator, as long as
@@ -152,9 +156,10 @@ class Instruction(BaseModel):
     CC_PRX
     ------
 
-    Classically controlled PRX gate. Takes three arguments. ``angle_t`` and ``phase_t`` are exactly as in PRX.
-    ``feedback_label == f"{physical_qubit_name}__{feedback_key}`` is a string that identifies the
-    ``measure`` operation and the qubit within it whose measurement result controls the gate.
+    Classically controlled PRX gate. Takes four arguments. ``angle_t`` and ``phase_t`` are exactly as in PRX.
+    ``feedback_key`` is a string that identifies the ``measure`` instruction whose result controls
+    the gate (the one that shares the feedback key).
+    ``feedback_qubit`` is the name of the physical qubit within the ``measure`` instruction that produces the feedback.
     If the measurement result is 1, the PRX gate is applied. If it is 0, an identity gate of similar time
     duration gate is applied instead.
     The measurement instruction must precede the classically controlled gate instruction in the quantum circuit.
@@ -517,6 +522,14 @@ class QuantumArchitecture(BaseModel):
     """Details about the quantum architecture."""
 
 
+def _component_sort_key(component_name: str) -> tuple[str, int, str]:
+    def get_numeric_id(name: str) -> int:
+        match = re.search(r'(\d+)', name)
+        return int(match.group(1)) if match else 0
+
+    return re.sub(r'[^a-zA-Z]', '', component_name), get_numeric_id(component_name), component_name
+
+
 class GateImplementationInfo(BaseModel):
     """Information about an implementation of a quantum gate/operation."""
 
@@ -536,6 +549,28 @@ class GateInfo(BaseModel):
     override_default_implementation: dict[Locus, str] = Field(...)
     """mapping of loci to implementation names that override ``default_implementation`` for those loci"""
 
+    @cached_property
+    def loci(self) -> tuple[Locus, ...]:
+        """Returns all loci which are available for at least one of the implementations.
+
+        The loci are sorted first based on the first locus component, then the second, etc.
+        The sorting of individual locus components is first done alphabetically based on their
+        non-numeric part, and then components with the same non-numeric part are sorted numerically.
+        An example of loci sorted this way would be:
+        (
+            ('QB1', 'QB2'),
+            ('QB2', 'COMPR1'),
+            ('QB2', 'QB3'),
+            ('QB3', 'COMPR1'),
+            ('QB3', 'COMPR2'),
+            ('QB3', 'QB1'),
+            ('QB10', 'QB2'),
+        )
+        """
+        loci_set = set(locus for impl in self.implementations.values() for locus in impl.loci)
+        loci_sorted = sorted(loci_set, key=lambda locus: tuple(map(_component_sort_key, locus)))
+        return tuple(loci_sorted)
+
 
 class DynamicQuantumArchitecture(BaseModel):
     """Dynamic quantum architecture as returned by server.
@@ -552,6 +587,16 @@ class DynamicQuantumArchitecture(BaseModel):
     """computational resonators that appear in at least one gate locus in the calibration set"""
     gates: dict[str, GateInfo] = Field(...)
     """mapping of gate names to information about the gates"""
+
+    @cached_property
+    def components(self) -> tuple[str, ...]:
+        """Returns all locus components (qubits and computational resonators) sorted.
+
+        The components are first sorted alphabetically based on their non-numeric part, and then
+        components with the same non-numeric part are sorted numerically. An example of components
+        sorted this way would be: ('COMPR1', 'COMPR2', 'QB1', 'QB2', 'QB3', 'QB10', 'QB11', 'QB20').
+        """
+        return tuple(sorted(self.qubits + self.computational_resonators, key=_component_sort_key))
 
 
 class HeraldingMode(str, Enum):